add generic event
